Better progress for all types
Introducing a trickle system, which works better for UIWebView for the resource loading part.

The King has reduced responsibility and maybe can be removed.

This system should make it easier to make the SuProgressBar handle adding new progress types during existing monitoring operations, which seems necessary for this to be a good library.